backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.129.0",
    "uvicorn>=0.41.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "moviepy>=1.0.3",
    "edge-tts>=6.1.12",
    "google-generativeai>=0.8.0",
    "google-genai>=1.64.0",
    "elevenlabs>=2.36.1",
    "supabase>=2.28.0",
]
